+++ backend/src/status/model.rs
@@ -3,8 +3,8 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize)]
 pub struct Status {
   pub version: String,
-  pub max_size: usize,
-  pub max_views: usize,
-  pub max_expiration: usize,
+  pub max_size: u32,
+  pub max_views: u32,
+  pub max_expiration: u32,
   pub allow_advanced: bool,
 }
